Add initial Task component with atomic architecture
I've chosen atomic architecture because the project size works well with it, but I have not nested directories due to the small size.

I used A11yproject.com to check colour contrast on the text and used css variables, but they could be better refactored to come from the project root, or for the task-priority-specific ones to be determined in the Task.css file and not the child component files.