Adds TaskEdit Component, reducer, and update functionality
This is far more than should be a in a single commit, but I haven't been building in nice, modular, task-planned pieces of work.

I used a reducer because I had some issues with updating state in the update component; on reflection they may have been caused by bugs elsewhere but it's a very nice pattern to demonstrate an understanding on redux et al.